URLStorage: if the download fails (or the response read fails), raise a spacer.exceptions.URLDownloadError.
This replaces SpacerInputError for download failures. The concept of SpacerInputError was flawed because, in general, pyspacer should not be able to tell which errors are the end user's "fault" or not. That depends on the application.

With URLDownloadError, pyspacer is taking up the job of grouping certain errors usefully; this tells the error handler (in the calling program, not in pyspacer) whether the download/read error was associated with a URLStorage instance or something else. But it's up to the handler to decide the error's root cause and the appropriate handling action.