Move VGG16CaffeExtractor to extractors/vgg16.py, and collapse helper functions similarly to EfficientNet.
Collapsing helper functions seemed worth it because there was a lot of parameter passing compared to the size of the rest of the code.

load_net() is still separate so that its @lru_cache can have maximum effect.

The biggest goal of this refactor was removing VGG16/Caffe specifics from extractors/base.py.