torch >=2.6,<2.9 and torchvision >=0.21,<0.24. This adds lenience in both the new and old directions.
torch 2.8 is currently the minimum to address GitHub security alerts, but torch is a big package to pin to a very recent version.

For example, SageMaker's default torch version is 2.6 at this time of writing, and the time needed to install a different torch version can make it onerous to use a newer torch version with SageMaker.