Train message: have the ImageLabels parameter specify all feature-location info, not just the file paths.
Now that the storage type and bucket name aren't specified by a separate features_loc parameter, it's possible for training data to come from multiple buckets or storage types.

Meanwhile, ImageLabels now encapsulates more data-loading details. This includes the filesystem_cache which used to be specified on DataLocation, and the batching logic which used to be in loose functions of train_utils.py.